--- src/services/middlewares/bodyparser.js
@@ -1,6 +1,0 @@
-import bodyParser from 'body-parser';
-
-export default app => app.use(
-    bodyParser.json(),
-    bodyParser.urlencoded({ extended: false })
-);
